+++ src/main/java/com/khronodragon/bluestone/enums/MessageDestination.java
@@ -4,16 +4,16 @@ import com.khronodragon.bluestone.Context;
 import net.dv8tion.jda.core.entities.MessageChannel;
 
 public enum MessageDestination {
-    CHANNEL((short)1),
-    AUTHOR((short)2);
+    CHANNEL((byte)1),
+    AUTHOR((byte)2);
 
-    private final short type;
-    MessageDestination(short type) {
+    private final byte type;
+    MessageDestination(byte type) {
         this.type = type;
     }
 
     public MessageChannel getChannel(Context ctx) {
-        if (type == (short)1) {
+        if (type == (byte)1) {
             return ctx.channel;
         } else {
             return ctx.author.openPrivateChannel().complete();
